+++ src/ai-bundle/tests/DependencyInjection/AiBundleTest.php
@@ -120,6 +120,13 @@ class AiBundleTest extends TestCase
         $this->assertArrayHasKey('console.command', $dropStoreCommandDefinition->getTags());
     }
 
+    public function testMessageBagNormalizerIsRegistered()
+    {
+        $container = $this->buildContainer($this->getFullConfig());
+
+        $this->assertTrue($container->hasDefinition('ai.chat.message_bag.normalizer'));
+    }
+
     public function testInjectionAgentAliasIsRegistered()
     {
         $container = $this->buildContainer([
@@ -3040,6 +3047,15 @@ class AiBundleTest extends TestCase
                             'endpoint' => 'http://127.0.0.1:7700',
                             'api_key' => 'foo',
                             'index_name' => 'test',
+                        ],
+                    ],
+                    'redis' => [
+                        'my_redis_store' => [
+                            'connection_parameters' => [
+                                'host' => '1.2.3.4',
+                                'port' => 6379,
+                            ],
+                            'index_name' => 'my_message_store',
                         ],
                     ],
                     'session' => [